Replace Jest with Node built-in test runner
