+++ package.json
@@ -80,21 +80,18 @@
     "@babel/cli": "^7.23.4",
     "@babel/core": "^7.23.2",
     "@babel/preset-env": "^7.23.2",
-    "babel-jest": "^29.7.0",
     "esbuild": "^0.19.9",
     "eslint": "^8.39.0",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-prettier": "^4.2.1",
     "fast-glob": "^3.2.12",
     "html_codesniffer": "^2.5.1",
-    "jest": "^29.5.0",
-    "jest-environment-jsdom": "^29.5.0",
-    "jest-puppeteer": "^9.0.0",
     "npm-run-all": "^4.1.5",
     "pixelmatch": "^5.3.0",
     "pngjs": "^7.0.0",
     "prettier": "^2.8.8",
     "puppeteer": "^20.8.0",
+    "serve-static": "^1.15.0",
     "stylelint": "^15.6.1",
     "svgo": "^3.0.2"
   }
