+++ test/.eslintrc.json
@@ -1,0 +1,6 @@
+{
+  "parserOptions": {
+    "ecmaVersion": "latest",
+    "sourceType": "module"
+  }
+}
