--- test/.eslintrc.yml
@@ -1,6 +1,0 @@
-env:
-  jest: true
-globals:
-  browser: true
-  page: true
-  context: true
